refactor(#218): split bootstrap logic in app.tsx
