+++ apps/frontend/src/App.tsx
@@ -4,33 +4,10 @@ import { Layout } from './components';
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 import { TooltipProvider } from '@/components';
 import { LandingPage, SignInPage, SignUpPage, CanvasPage } from '@/pages';
-import { useEffect } from 'react';
-import { authStore } from '@/store/auth.store';
-import { getMyInfo, refreshToken } from '@/lib/api';
-import { RequireAuth } from '@/features/auth';
+import { RequireAuth, useAuthBootstrap } from '@/features/auth';
 
 function App() {
-  useEffect(() => {
-    const bootstrapAuth = async () => {
-      try {
-        authStore.setAuthLoading(true);
-        await refreshToken();
-
-        const res = await getMyInfo();
-        if (res.success && res.result) {
-          authStore.setUser(res.result);
-        } else {
-          authStore.clearAuth();
-        }
-      } catch {
-        authStore.clearAuth();
-      } finally {
-        authStore.setAuthLoading(false);
-        authStore.setInitialized(true);
-      }
-    };
-    bootstrapAuth();
-  }, []);
+  useAuthBootstrap();
 
   return (
     <ThemeProvider defaultTheme="system" storageKey="schemafy-theme">
